packages/uvicorn-hmr/pyproject.toml
[project]
name = "uvicorn-hmr"
description = "Hot Module Reloading for Uvicorn"
version = "0.1.1.8"
readme = "README.md"
requires-python = ">=3.12"
keywords = ["uvicorn", "hot-reload", "hmr", "reload", "server"]
authors = [{ name = "Muspi Merol", email = "me@promplate.dev" }]
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
]
dependencies = [
    "hmr~=0.7.0",
    "hmr-runner>=0.1.0,<1",
    "typer-slim>=0.15.4,<1",
    "uvicorn>=0.24.0",
]

[project.scripts]
uvicorn-hmr = "uvicorn_hmr:app"

[project.urls]
Homepage = "https://github.com/promplate/hmr"
Documentation = "https://hmr.promplate.dev"
Repository = "https://github.com/promplate/hmr/tree/HEAD/packages/uvicorn-hmr"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.optional-dependencies]
all = [
    "fastapi-reloader~=1.0",
]

[tool.uv.sources]
fastapi-reloader = { workspace = true }
hmr-runner = { workspace = true }
